requirements.txt
Flask==0.12
Flask-Migrate==1.8.0
Flask-Cors=3.0.2
SQLAlchemy==1.1.5
psycopg2==2.6.2
python-dateutil==2.6.0
pytz==2016.10
